Remove debug output
Now, it is v0.1! The debugged output - which is a development thing - is only enabled if the 2nd argument (and only the second argument) is '-d'. This is a hidden argument used only for development purposes. As of official v0.1, it's not important to the final user to see the hex data related to the middle of conversion algorithm.